move functions to app_manager/ui_translations
