+++ corehq/apps/app_manager/ui_translations/__init__.py
@@ -1,0 +1,5 @@
+from .ui_translations import (
+    process_ui_translation_upload,
+    build_ui_translation_download_file,
+    get_default_translations_for_download,
+)
